AG-10004 - Move thumbnail generation code.
