+++ plugins/ag-charts-build-tools/src/executors/generate-chart-thumbnail/generator/jsCodeShiftUtils.ts
