+++ plugins/ag-charts-build-tools/src/executors/generate-chart-thumbnail/generator/transformPlainEntryFile.test.ts
