+++ plugins/ag-charts-build-tools/src/executors/generate-chart-thumbnail/generator/transformPlainEntryFile.ts
